- minor fix : changed float types to CGFloat. Necessary for 64-bit Cocoa transition.

+++ Colorboard/BNRColorViewController.m
@@ -39,7 +39,7 @@
     UIColor *color = self.colorDescription.color;
     
     // Get the RGB values out of the UIColor object
-    float red, green, blue;
+    CGFloat red, green, blue;
     [color getRed:&red green:&green blue:&blue alpha:nil];
     
     // Set the initial slider values
